+++ src/components/stack.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 void init_stack();
-void stack_push(uint8_t value);
+void stack_push(uint16_t value);
 uint16_t stack_pop();
 
 #endif //CHIP_8_EMULATOR_STACK_H
